moved the AetherPlugin back to compat
